+++ bpmn-examples/src/main/java/org/bpmnwithactiviti/chapter12/ruletask/ApplicationCreator.java
@@ -9,8 +9,8 @@ public class ApplicationCreator implements JavaDelegate {
 	public void execute(DelegateExecution execution) {
 		LoanApplicant applicant = new LoanApplicant();
 		applicant.setName((String) execution.getVariable("name"));
-		applicant.setIncome((Integer) execution.getVariable("income"));
-		applicant.setLoanAmount((Integer) execution.getVariable("loanAmount"));
+		applicant.setIncome((Long) execution.getVariable("income"));
+		applicant.setLoanAmount((Long) execution.getVariable("loanAmount"));
 		applicant.setEmailAddress((String) execution.getVariable("emailAddress"));
 		execution.setVariable("loanApplicant", applicant);
 	}
